5. We need to have a look at the features of roles
User specifications

User must be able to View Devices and Customers (No privilege for adding
or updating)
Must be able to receive spares.
The user must be able to must Make order.
The user must be able to Make a leave.
If part number does not exist on receive spares there must not be an
option add a new part number.
User must be able to create order on behalf of the technician but no
option to see other options such as approved orders, received,orders to
receive. The user must be able to such by technician and customer.